+++ jobs/pacman-awsrules/src/main/java/com/tmobile/cloud/contrastrules/library/LibraryVulnerabilityAssessmentPolicy.java
@@ -20,11 +20,12 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.nimbusds.oauth2.sdk.util.MapUtils;
 import com.tmobile.cloud.awsrules.utils.PacmanUtils;
 import com.tmobile.cloud.constants.PacmanRuleConstants;
-import com.tmobile.cloud.contrastrules.utils.ResourceVulnerability;
+import com.tmobile.cloud.contrastrules.model.ResourceVulnerability;
 import com.tmobile.cloud.model.CveDetails;
 import com.tmobile.cloud.model.VulnerabilityInfo;
 import com.tmobile.pacman.commons.PacmanSdkConstants;
 import com.tmobile.pacman.commons.exception.InvalidInputException;
+import com.tmobile.pacman.commons.exception.RuleExecutionFailedExeption;
 import com.tmobile.pacman.commons.policy.Annotation;
 import com.tmobile.pacman.commons.policy.BasePolicy;
 import com.tmobile.pacman.commons.policy.PacmanPolicy;
@@ -48,6 +49,7 @@ public class LibraryVulnerabilityAssessmentPolicy extends BasePolicy {
 
     private static final String CONTRAST_URL_TEMPLATE = "https://eval.contrastsecurity.com/Contrast/static/ng/" +
             "index.html#/%s/libraries?quickFilter=VULNERABLE&filterText=%s";
+    private static final ObjectMapper mapper = new ObjectMapper();
 
     /**
      * @param policyParam        parameters for policy validation
@@ -56,8 +58,9 @@ public class LibraryVulnerabilityAssessmentPolicy extends BasePolicy {
     @Override
     public PolicyResult execute(final Map<String, String> policyParam, Map<String, String> resourceAttributes) {
         logger.debug("LibraryVulnerabilityAssessmentPolicy started");
-        if (MapUtils.isNotEmpty(policyParam) && !PacmanUtils.doesAllHaveValue(policyParam.get(PacmanRuleConstants.SEVERITY),
-                policyParam.get(PacmanRuleConstants.CATEGORY))) {
+        if (MapUtils.isEmpty(policyParam) || MapUtils.isEmpty(resourceAttributes) || (MapUtils.isNotEmpty(policyParam)
+                && !PacmanUtils.doesAllHaveValue(policyParam.get(PacmanRuleConstants.SEVERITY),
+                policyParam.get(PacmanRuleConstants.CATEGORY)))) {
             logger.info(PacmanRuleConstants.MISSING_CONFIGURATION);
             throw new InvalidInputException(PacmanRuleConstants.MISSING_CONFIGURATION);
         }
@@ -65,12 +68,10 @@ public class LibraryVulnerabilityAssessmentPolicy extends BasePolicy {
         String accountId = resourceAttributes.get(PacmanRuleConstants.ACCOUNTID);
         MDC.put("executionId", policyParam.get("executionId"));
         MDC.put("policyId", policyId);
-        String vulnerabilities;
         String severity = policyParam.get("severityToCheck");
         List<VulnerabilityInfo> vulnerabilityDetailsList = new ArrayList<>();
-        ObjectMapper mapper = new ObjectMapper();
         try {
-            vulnerabilities = resourceAttributes.get("vulns");
+            String vulnerabilities = resourceAttributes.get("vulns");
             if (vulnerabilities == null || vulnerabilities.isEmpty()) {
                 return new PolicyResult(PacmanSdkConstants.STATUS_SUCCESS, PacmanRuleConstants.SUCCESS_MESSAGE);
             }
@@ -88,7 +89,7 @@ public class LibraryVulnerabilityAssessmentPolicy extends BasePolicy {
                 List<CveDetails> cveList = new ArrayList<>();
                 VulnerabilityInfo vulnerability = new VulnerabilityInfo();
                 vulnerability.setVulnerabilityUrl(vulnerabilityUrl);
-                vulnerability.setTitle(vul.getDescription().replace("\u00a0",""));
+                vulnerability.setTitle(vul.getDescription().replace("\u00a0", ""));
                 CveDetails cve = new CveDetails();
                 cve.setId(vul.getName());
                 cve.setUrl(PacmanUtils.NIST_VULN_DETAILS_URL + vul.getName());
@@ -99,8 +100,8 @@ public class LibraryVulnerabilityAssessmentPolicy extends BasePolicy {
             return buildFailureAnnotation(policyParam, vulnerabilityDetailsList,
                     resourceAttributes);
         } catch (Exception e) {
-            logger.error(e.getMessage(), e);
-            return buildFailureAnnotation(policyParam, vulnerabilityDetailsList, resourceAttributes);
+            logger.error("Unable to get vulnerability details", e);
+            throw new RuleExecutionFailedExeption(e.getMessage());
         }
     }
 
@@ -112,16 +113,16 @@ public class LibraryVulnerabilityAssessmentPolicy extends BasePolicy {
     private PolicyResult buildFailureAnnotation(final Map<String, String> policyParam,
                                                 List<VulnerabilityInfo> vulnerabilityDetailsList,
                                                 Map<String, String> resourceAttributes) {
-        ObjectMapper mapper = new ObjectMapper();
+        int vulCount = vulnerabilityDetailsList.size();
         String violationDetails = "";
         String fileName = resourceAttributes.get("fileName");
         try {
             violationDetails = mapper.writeValueAsString(vulnerabilityDetailsList);
         } catch (Exception e) {
-            logger.error("error while converting violation details for {}", fileName);
+            logger.error(e.getMessage(), e);
         }
         String policyId = policyParam.get(PacmanSdkConstants.POLICY_ID);
-        String desc = String.format("%s has %d vulnerabilities", fileName, vulnerabilityDetailsList.size());
+        String desc = fileName + " has " + vulCount + (vulCount == 1 ? " vulnerability" : " vulnerabilities");
         LinkedHashMap<String, Object> issue = new LinkedHashMap<>();
         List<LinkedHashMap<String, Object>> issueList = new ArrayList<>();
         Annotation annotation = Annotation.buildAnnotation(policyParam, Annotation.Type.ISSUE);
