+++ jobs/pacman-awsrules/src/main/java/com/tmobile/cloud/contrastrules/model/ResourceVulnerability.java
@@ -1,0 +1,49 @@
+/*******************************************************************************
+ *  Copyright 2023 Paladin Cloud, Inc. or its affiliates. All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ *  use this file except in compliance with the License.  You may obtain a copy
+ *  of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ *  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.  See the
+ *  License for the specific language governing permissions and limitations under
+ *  the License.
+ ******************************************************************************/
+package com.tmobile.cloud.contrastrules.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class ResourceVulnerability {
+    public String name;
+    public String description;
+    public String severityToUse;
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public String getDescription() {
+        return description;
+    }
+
+    public void setDescription(String description) {
+        this.description = description;
+    }
+
+    public String getSeverityToUse() {
+        return severityToUse;
+    }
+
+    public void setSeverityToUse(String severityToUse) {
+        this.severityToUse = severityToUse;
+    }
+}
