--- jobs/pacman-awsrules/src/main/java/com/tmobile/cloud/contrastrules/utils/ResourceVulnerability.java
@@ -1,34 +1,0 @@
-package com.tmobile.cloud.contrastrules.utils;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class ResourceVulnerability {
-    public String name;
-    public String description;
-    public String severityToUse;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public String getSeverityToUse() {
-        return severityToUse;
-    }
-
-    public void setSeverityToUse(String severityToUse) {
-        this.severityToUse = severityToUse;
-    }
-}
